Extract the Typst stuff out of utils, update typst to 0.14.0, remove useless UI elements
