+++ Cargo.toml
@@ -81,9 +81,9 @@ svg = "0.18.0"
 thiserror = "2.0.12"
 tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
-typst = "0.12.0"
-typst-kit = { version = "0.12.0", features = ["embed-fonts"] }
-typst-svg = "0.12.0"
+typst = "0.14.0"
+typst-kit = { version = "0.14.0", features = ["embed-fonts"] }
+typst-svg = "0.14.0"
 unicode-segmentation = "1.12"
 url = "2.5"
 usvg = "0.45.1"
